refactor model directory structure
