--- design-parking-lot/src/main/java/com/sanket/designparkinglot/models/GateStatus.java
@@ -1,6 +1,0 @@
-package com.sanket.designparkinglot.models;
-
-public enum GateStatus {
-    OPEN,
-    CLOSED
-}
