--- design-parking-lot/src/main/java/com/sanket/designparkinglot/models/GateType.java
@@ -1,6 +1,0 @@
-package com.sanket.designparkinglot.models;
-
-public enum GateType {
-    ENTRY,
-    EXIT
-}
