--- design-parking-lot/src/main/java/com/sanket/designparkinglot/models/ParkingLot.java
@@ -1,17 +1,0 @@
-package com.sanket.designparkinglot.models;
-
-import lombok.Getter;
-import lombok.Setter;
-
-import java.util.List;
-
-@Getter
-@Setter
-public class ParkingLot extends BaseModel {
-
-    private String address;
-    private List<Floor> floors;
-    private List<EntryGate> entryGates;
-    private List<ExitGate> exitGates;
-
-}
