+++ design-parking-lot/src/main/java/com/sanket/designparkinglot/models/bill/Bill.java
@@ -1,5 +1,9 @@
-package com.sanket.designparkinglot.models;
+package com.sanket.designparkinglot.models.bill;
 
+import com.sanket.designparkinglot.models.BaseModel;
+import com.sanket.designparkinglot.models.operator.Operator;
+import com.sanket.designparkinglot.models.ticket.Ticket;
+import com.sanket.designparkinglot.models.gates.ExitGate;
 import lombok.Getter;
 import lombok.Setter;
