+++ design-parking-lot/src/main/java/com/sanket/designparkinglot/models/floor/Floor.java
@@ -1,5 +1,7 @@
-package com.sanket.designparkinglot.models;
+package com.sanket.designparkinglot.models.floor;
 
+import com.sanket.designparkinglot.models.BaseModel;
+import com.sanket.designparkinglot.models.spot.Spot;
 import lombok.Getter;
 import lombok.Setter;
