+++ design-parking-lot/src/main/java/com/sanket/designparkinglot/models/gates/GateStatus.java
@@ -1,0 +1,6 @@
+package com.sanket.designparkinglot.models.gates;
+
+public enum GateStatus {
+    OPEN,
+    CLOSED
+}
