+++ design-parking-lot/src/main/java/com/sanket/designparkinglot/models/gates/GateType.java
@@ -1,0 +1,6 @@
+package com.sanket.designparkinglot.models.gates;
+
+public enum GateType {
+    ENTRY,
+    EXIT
+}
