+++ design-parking-lot/src/main/java/com/sanket/designparkinglot/models/operator/Operator.java
@@ -1,5 +1,6 @@
-package com.sanket.designparkinglot.models;
+package com.sanket.designparkinglot.models.operator;
 
+import com.sanket.designparkinglot.models.BaseModel;
 import lombok.Getter;
 import lombok.Setter;
