+++ design-parking-lot/src/main/java/com/sanket/designparkinglot/models/payment/Payment.java
@@ -1,5 +1,7 @@
-package com.sanket.designparkinglot.models;
+package com.sanket.designparkinglot.models.payment;
 
+import com.sanket.designparkinglot.models.BaseModel;
+import com.sanket.designparkinglot.models.bill.Bill;
 import lombok.Getter;
 import lombok.Setter;
