+++ design-parking-lot/src/main/java/com/sanket/designparkinglot/models/payment/PaymentMode.java
@@ -1,4 +1,4 @@
-package com.sanket.designparkinglot.models;
+package com.sanket.designparkinglot.models.payment;
 
 public enum PaymentMode {
     CASH,
