+++ design-parking-lot/src/main/java/com/sanket/designparkinglot/models/spot/SpotStatus.java
@@ -1,4 +1,4 @@
-package com.sanket.designparkinglot.models;
+package com.sanket.designparkinglot.models.spot;
 
 public enum SpotStatus {
     AVAILABLE,
