+++ design-parking-lot/src/main/java/com/sanket/designparkinglot/models/spot/SpotType.java
@@ -1,4 +1,4 @@
-package com.sanket.designparkinglot.models;
+package com.sanket.designparkinglot.models.spot;
 
 public enum SpotType {
     BIKE,
