+++ design-parking-lot/src/main/java/com/sanket/designparkinglot/models/vehicle/VehicleType.java
@@ -1,4 +1,4 @@
-package com.sanket.designparkinglot.models;
+package com.sanket.designparkinglot.models.vehicle;
 
 public enum VehicleType {
     BIKE,
