+++ design-parking-lot/src/main/java/com/sanket/designparkinglot/strategies/feescalculator/FeesCalculatorStrategy.java
@@ -1,6 +1,6 @@
 package com.sanket.designparkinglot.strategies.feescalculator;
 
-import com.sanket.designparkinglot.models.Ticket;
+import com.sanket.designparkinglot.models.ticket.Ticket;
 
 public interface FeesCalculatorStrategy {
