+++ design-parking-lot/src/main/java/com/sanket/designparkinglot/strategies/spotassignment/RandomSpotAssignmentStrategy.java
@@ -1,7 +1,12 @@
 package com.sanket.designparkinglot.strategies.spotassignment;
 
 import com.sanket.designparkinglot.exceptions.NoSpotAvailableException;
-import com.sanket.designparkinglot.models.*;
+import com.sanket.designparkinglot.models.floor.Floor;
+import com.sanket.designparkinglot.models.gates.EntryGate;
+import com.sanket.designparkinglot.models.parkinglot.ParkingLot;
+import com.sanket.designparkinglot.models.spot.Spot;
+import com.sanket.designparkinglot.models.spot.SpotStatus;
+import com.sanket.designparkinglot.models.vehicle.Vehicle;
 
 import java.util.List;
