+++ design-parking-lot/src/main/java/com/sanket/designparkinglot/strategies/spotassignment/SpotAssignmentStrategy.java
@@ -1,10 +1,10 @@
 package com.sanket.designparkinglot.strategies.spotassignment;
 
 import com.sanket.designparkinglot.exceptions.NoSpotAvailableException;
-import com.sanket.designparkinglot.models.EntryGate;
-import com.sanket.designparkinglot.models.ParkingLot;
-import com.sanket.designparkinglot.models.Spot;
-import com.sanket.designparkinglot.models.Vehicle;
+import com.sanket.designparkinglot.models.gates.EntryGate;
+import com.sanket.designparkinglot.models.parkinglot.ParkingLot;
+import com.sanket.designparkinglot.models.spot.Spot;
+import com.sanket.designparkinglot.models.vehicle.Vehicle;
 
 public interface SpotAssignmentStrategy {
